+++ flexiapi/resources/views/mails/register_confirmation_text.blade.php
@@ -1,9 +1,0 @@
-Hello,
-
-You just created an account on {{ space()->name }} using your email account.
-Please follow the unique link bellow to set up your password and finish the registration process.
-
-{{ $link }}
-
-Regards,
-{{ config('mail.signature') }}